Move load_spatialite and improve the doc for Alembic
